Feat)advice

+++ src/main/java/com/example/todo/model/UserEntity.java
@@ -41,7 +41,7 @@ public class UserEntity {
 	@Column(nullable=false)
 	private String password;
 
-	private Integer height;
-	private Integer weight;
+	private Double height;
+	private Double weight;
 
 }
